+++ core-config-fe/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @groovestack/versions
 
+## 0.2.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/base@0.2.13
+
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/base@0.2.12
+
 ## 0.2.11
 
 ### Patch Changes