+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # docs
 
+## 0.0.4
+
+### Patch Changes
+
+- StatusInput gracefully handles undefined save context
+
+## 0.0.3
+
+### Patch Changes
+
+- useSaveOnlyChanges & enhance StatusInput
+
 ## 0.0.2
 
 ### Patch Changes